requirements.txt
Pillow>=10,<13
requests>=2.32,<3
pyperclip>=1.9,<2
keyboard>=0.13.5,<1
pystray>=0.19,<1
plyer>=2.1,<3
